refactor(explorer): split the 799-line god-object; shared panel kit
Behavior-preserving extraction (same DOM, CSS classes, aria, wiring):

- src/views/explorer/ gains hero-card.ts (cover mgmt + inline target),
  ideas-inbox.ts, reconcile-banner.ts (Tier-2 relink UI), scene-rows.ts
  (badges/drag/indent/menus), series-menu.ts; explorer-view.ts is now a
  232-line orchestrator.
- src/views/panel-kit.ts: renderEmptyState (replaces the hand-rolled
  empty-state div in write/beat/board/structure panels) and SectionState
  (audit-panel's details-open-state preservation, generalized; audit-panel
  adopts it).

Verified live: Home hero/scene rows/badges and the 16 Audit sections all
render with no errors.

[skip changelog]
